restructured servlet packages
