+++ galasa-parent/dev.galasa.framework.api.ras/src/main/java/dev/galasa/framework/api/ras/internal/commons/IRunRootArtifact.java
@@ -1,14 +1,14 @@
 /*
  * Copyright contributors to the Galasa project 
  */
-package dev.galasa.framework.api.ras.internal;
+package dev.galasa.framework.api.ras.internal.commons;
 
 import java.io.IOException;
 
 import dev.galasa.framework.spi.IRunResult;
 import dev.galasa.framework.spi.ResultArchiveStoreException;
 
-interface IRunRootArtifact {
+public interface IRunRootArtifact {
     String getPathName();
     byte[] getContent(IRunResult run) throws ResultArchiveStoreException, IOException;
     String getContentType();
