+++ galasa-parent/dev.galasa.framework.api.ras/src/main/java/dev/galasa/framework/api/ras/internal/routes/RunArtifactsDownloadRoute.java
@@ -1,7 +1,9 @@
 /*
  * Copyright contributors to the Galasa project
  */
-package dev.galasa.framework.api.ras.internal;
+package dev.galasa.framework.api.ras.internal.routes;
+
+import static dev.galasa.framework.api.ras.internal.commons.ServletErrorMessage.*;
 
 import java.io.IOException;
 import java.io.OutputStream;
@@ -25,13 +27,18 @@ import javax.servlet.http.HttpServletResponse;
 import com.google.gson.Gson;
 
 import dev.galasa.framework.IFileSystem;
+import dev.galasa.framework.api.ras.internal.commons.ArtifactPropertiesArtifact;
+import dev.galasa.framework.api.ras.internal.commons.IRunRootArtifact;
+import dev.galasa.framework.api.ras.internal.commons.InternalServletException;
+import dev.galasa.framework.api.ras.internal.commons.QueryParameters;
+import dev.galasa.framework.api.ras.internal.commons.RunLogArtifact;
+import dev.galasa.framework.api.ras.internal.commons.ServletError;
+import dev.galasa.framework.api.ras.internal.commons.StructureJsonArtifact;
 import dev.galasa.framework.spi.FrameworkException;
 import dev.galasa.framework.spi.IFramework;
 import dev.galasa.framework.spi.IRunResult;
 import dev.galasa.framework.spi.ResultArchiveStoreException;
 import dev.galasa.framework.spi.utils.GalasaGsonBuilder;
-
-import static dev.galasa.framework.api.ras.internal.ServletErrorMessage.*;
 
 /**
  * Implementation to download an artifact for a given run based on its runId and the path
