+++ galasa-parent/dev.galasa.framework.api.ras/src/main/java/dev/galasa/framework/api/ras/internal/routes/RunArtifactsListRoute.java
@@ -1,7 +1,7 @@
 /*
  * Copyright contributors to the Galasa project 
  */
-package dev.galasa.framework.api.ras.internal;
+package dev.galasa.framework.api.ras.internal.routes;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -17,14 +17,21 @@ import com.google.gson.Gson;
 import com.google.gson.JsonArray;
 
 import dev.galasa.framework.IFileSystem;
+import dev.galasa.framework.api.ras.internal.commons.ArtifactPropertiesArtifact;
+import dev.galasa.framework.api.ras.internal.commons.IRunRootArtifact;
+import dev.galasa.framework.api.ras.internal.commons.InternalServletException;
+import dev.galasa.framework.api.ras.internal.commons.QueryParameters;
+import dev.galasa.framework.api.ras.internal.commons.RunLogArtifact;
+import dev.galasa.framework.api.ras.internal.commons.ServletError;
+import dev.galasa.framework.api.ras.internal.commons.StructureJsonArtifact;
 import dev.galasa.framework.spi.FrameworkException;
 import dev.galasa.framework.spi.IFramework;
 import dev.galasa.framework.spi.IRunResult;
 import dev.galasa.framework.spi.ResultArchiveStoreException;
 import dev.galasa.framework.spi.utils.GalasaGsonBuilder;
 
-import static dev.galasa.framework.api.ras.internal.ServletErrorMessage.*;
 import static dev.galasa.framework.api.ras.internal.BaseServlet.*;
+import static dev.galasa.framework.api.ras.internal.commons.ServletErrorMessage.*;
 
 /**
  * Implementation to retrieve a list of artifacts for a given run based on its runId.
